move ProjectCard component into Project
